backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.3
qdrant-client==1.7.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pypdf==4.0.2
python-dotenv==1.0.1